+++ examples/code_tasks/tiny_digits_mlp_project/digits_mlp/__init__.py
@@ -1,0 +1,10 @@
+from digits_mlp.data import load_digits_split
+from digits_mlp.model import TinyMLP
+from digits_mlp.train import BenchmarkConfig, run_benchmark
+
+__all__ = [
+    "BenchmarkConfig",
+    "TinyMLP",
+    "load_digits_split",
+    "run_benchmark",
+]